refactor: unify STREAM_MESSAGE_TYPES imports to streaming-types
The previous circular-import fix left the codebase with two valid
import paths for the same constant: src/api/streaming (via a re-export)
and src/api/streaming-types (the canonical source). Two paths drift.

Migrate the remaining four consumers (IncomingMsg, MessagesContext,
useStreamingHandler, responseParser) to streaming-types and drop the
re-export from streaming.ts. One source, one path.
